+++ src/shared/ui/button/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Button"
