+++ src/shared/ui/card/index.tsx
@@ -1,0 +1,4 @@
+export * from "./Card"
+export * from "./CardContent"
+export * from "./CardHeader"
+export * from "./CardTitle"
