+++ src/shared/ui/dialog/DialogHeader.tsx
@@ -1,0 +1,11 @@
+import { forwardRef } from "react"
+
+interface DialogHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
+
+export const DialogHeader = forwardRef<HTMLDivElement, DialogHeaderProps>(({ className = "", ...props }, ref) => (
+  <div className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`} {...props} ref={ref} />
+))
+
+DialogHeader.displayName = "DialogHeader"
