+++ src/shared/ui/dialog/DialogOverlay.tsx
@@ -1,0 +1,3 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+export const DialogOverlay = DialogPrimitive.Overlay
