+++ src/shared/ui/form/index.tsx
@@ -1,0 +1,2 @@
+export * from "./Input"
+export * from "./Textarea"
