+++ src/shared/ui/select/index.tsx
@@ -1,0 +1,6 @@
+export * from "./Select"
+export * from "./SelectContent"
+export * from "./SelectGroup"
+export * from "./SelectItem"
+export * from "./SelectTrigger"
+export * from "./SelectValue"
