+++ src/shared/ui/table/TableHeader.tsx
@@ -1,0 +1,13 @@
+import { forwardRef } from "react"
+
+interface TableHeaderProps extends React.TableHTMLAttributes<HTMLTableSectionElement> {
+  className?: string
+}
+
+export const TableHeader = forwardRef<HTMLTableSectionElement, TableHeaderProps>(
+  ({ className = "", ...props }, ref) => {
+    return <thead ref={ref} className={`[&_tr]:border-b ${className}`} {...props} />
+  },
+)
+
+TableHeader.displayName = "TableHeader"
