+++ src/shared/ui/table/TableRow.tsx
@@ -1,0 +1,17 @@
+import { forwardRef } from "react"
+
+interface TableRowProps extends React.TableHTMLAttributes<HTMLTableRowElement> {
+  className?: string
+}
+
+export const TableRow = forwardRef<HTMLTableRowElement, TableRowProps>(({ className = "", ...props }, ref) => {
+  return (
+    <tr
+      ref={ref}
+      className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
+      {...props}
+    />
+  )
+})
+
+TableRow.displayName = "TableRow"
